+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - #35 Fix error logging in the `Nest` channel
 - Add `publish/docker/latest` and `publish/docker/tag` make targets
 - #34 Add Buienradar weather description
+- #28 Prepare `Dockerfile` and `setup.py` for the `Ping` channel
 
 ### `0.8`