+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - #35 Continue troubleshooting the `Nest` channel
 - #35 Set connection timeout to 10 minutes
 - #37 Use a separate session for `Buienradar`
+- #37 Get rid of the global `ClientSession` instance
 
 ### `0.8`